Pull TreeView out as separate component
